refactor(front): changes Svgs to svgs
