+++ client/src/components/svgs/loader/index.ts
