+++ client/src/components/svgs/loader/loader.tsx
