+++ client/src/components/svgs/logo/index.ts
