+++ client/src/components/svgs/logo/logo-svg.tsx
