+++ client/src/pages/sign-in/sign-in.page.tsx
@@ -1,7 +1,7 @@
 import astronautImage from 'assets/images/astronaut-auth.png';
 import Container from 'components/container';
 import Link from 'components/link';
-import LogoSvg from 'components/Svgs/logo';
+import LogoSvg from 'components/svgs/logo';
 import { BLACK, WHITE } from 'styles/colors';
 import { twJoin } from 'tailwind-merge';
